+++ lib/shoulda/matchers/active_model/errors.rb
@@ -2,6 +2,7 @@ module Shoulda # :nodoc:
   module Matchers
     module ActiveModel # :nodoc:
       class CouldNotDetermineValueOutsideOfArray < RuntimeError; end
+      class NonNullableBooleanError < Shoulda::Matchers::Error; end
     end
   end
 end